Move to specifications utils
