+++ kratos/python/add_variable_utils_to_python.cpp
@@ -464,9 +464,6 @@ void AddVariableUtilsToPython(pybind11::module &m)
         .def("AddDof", &VariableUtils::AddDofWithReaction<Variable<double>>)
         .def_static("AddDofsList", &VariableUtils::AddDofsList)
         .def_static("AddDofsList", &VariableUtils::AddDofsWithReactionsList)
-        .def_static("GetDofsListFromSpecifications", &VariableUtils::GetDofsListFromSpecifications)
-        .def_static("GetDofsListFromElementsSpecifications", &VariableUtils::GetDofsListFromElementsSpecifications)
-        .def_static("GetDofsListFromConditionsSpecifications", &VariableUtils::GetDofsListFromConditionsSpecifications)
         .def("CheckVariableKeys", &VariableUtils::CheckVariableKeys)
         .def("UpdateCurrentToInitialConfiguration", &VariableUtils::UpdateCurrentToInitialConfiguration)
         .def("UpdateInitialToCurrentConfiguration", &VariableUtils::UpdateInitialToCurrentConfiguration)
