+++ kratos/python/add_variable_utils_to_python.h
@@ -17,7 +17,6 @@
 
 
 // System includes
-#include <pybind11/stl.h>
 #include <pybind11/pybind11.h>
 
 // External includes
