+++ kratos/tests/test_specifications_utilities.py
@@ -162,5 +162,25 @@ class TestSpecificationsUtilities(KratosUnittest.TestCase):
         docu = KratosMultiphysics.Parameters("""{"SurfaceLoadCondition3D3N"   : "This is a pure displacement condition"}""")
         self.assertEqual(KratosMultiphysics.SpecificationsUtilities.GetDocumention(model_part).IsEquivalentTo(docu), True)
 
+    def testGetDofsListFromSpecifications(self):
+        # Set the test model part
+        current_model = KratosMultiphysics.Model()
+        model_part = current_model.CreateModelPart("Main")
+        model_part.AddNodalSolutionStepVariable(KratosMultiphysics.DISTANCE)
+        model_part.CreateNewNode(1,0.0,0.0,0.0)
+        model_part.CreateNewNode(2,1.0,0.0,0.0)
+        model_part.CreateNewNode(3,0.0,1.0,0.0)
+        model_part.CreateNewNode(4,1.0,1.0,0.0)
+        prop_1 = model_part.CreateNewProperties(1)
+        model_part.CreateNewElement("DistanceCalculationElementSimplex2D3N",1,[1,2,3],prop_1)
+        model_part.CreateNewElement("DistanceCalculationElementSimplex2D3N",2,[2,4,3],prop_1)
+
+        # Get the DOFs list from the elements specifications
+        dofs_list = KratosMultiphysics.SpecificationsUtilities.GetDofsListFromSpecifications(model_part)
+
+        # Check the obtained DOFs list
+        expected_dofs_list = ["DISTANCE"]
+        self.assertEqual(dofs_list, expected_dofs_list)
+
 if __name__ == '__main__':
     KratosUnittest.main()
